requirements.txt
pylint
pytest
jupyter
click
pandas
sensible
ipython
flask
pylint-flask
flask-API
requests
flask
flasgger
pytest-flask
pytest-cov
matplotlib
seaborn
numpy
gunicorn
google-cloud==0.27.0
google-cloud-core==0.27.1
google-cloud-pubsub==0.28.3
google-cloud-datastore==1.3.0
google-cloud-spanner==0.28.0
google-cloud-storage==1.4.0
grpcio==1.27.2
